backend/requirements.txt
django
djangorestframework
django-cors-headers
pymongo
djongo
python-dotenv
